Weave shared assemblies
Updating shared assemblies using "bindingRedirect" works, but it forces a specific order of extraction for binaries. Wii U USB Helper would need to go in first, then USBHelperLauncher. This is problematic as it did not matter in the past. USBHelperInstaller does it in the opposite order.

With these changes, we now weave these shared assemblies into the binaries, avoiding any conflicts. There is a small size penalty to embedding Newtonsoft.Json in both assemblies separately. Ideally, they would be stored in a shared file.